replace all reference of float to float64 now works with last version of go

+++ draw2d/src/cmd/testandroid.go
@@ -39,7 +39,7 @@ func initGc(w, h int) (image.Image, *draw2d.GraphicContext) {
 
 func saveToPngFile(TestName string, m image.Image) {
 	dt := time.Nanoseconds() - lastTime
-	fmt.Printf("%s during: %f ms\n", TestName, float(dt)*10e-6)
+	fmt.Printf("%s during: %f ms\n", TestName, float64(dt)*10e-6)
 	filePath := folder + TestName + ".png"
 	f, err := os.Open(filePath, os.O_CREAT|os.O_WRONLY, 0600)
 	if err != nil {
@@ -61,7 +61,7 @@ func saveToPngFile(TestName string, m image.Image) {
 	fmt.Printf("Wrote %s OK.\n", filePath)
 }
 
-func android(gc *draw2d.GraphicContext, x, y float) {
+func android(gc *draw2d.GraphicContext, x, y float64) {
 	gc.SetLineCap(draw2d.RoundCap)
 	gc.SetLineWidth(5)
 	gc.ArcTo(x+80, y+70, 50, 50, 180*(math.Pi/180), 360*(math.Pi/180)) // head
